Channel islands import
